test(signal): isolate reply author state
