+++ extensions/signal/src/reply-authors.test-helpers.ts
@@ -1,0 +1,7 @@
+// Signal test support resets native-reply quote author state.
+import { signalReplyAuthorState } from "./reply-authors-state.js";
+
+export function resetSignalReplyAuthorsForTests(): void {
+  signalReplyAuthorState.memoryReplyContexts.clear();
+  signalReplyAuthorState.persistentStoreDisabled = false;
+}
